27: WIP:  Extracted to hook
