+++ src/components/SlidingCarousel/index.ts
@@ -1,2 +1,2 @@
 export { SlidingCarousel } from "./SlidingCarousel";
-export { SlidingCarouselProvider, calculateNearestSnapPoint, calculateScrollOffsetForIndex } from "./SlidingCarouselProvider";
+export { SlidingCarouselProvider } from "./SlidingCarouselProvider";
